backend/requirements.txt
django==6.0.3
djangorestframework==3.17.1
djangorestframework-simplejwt==5.5.1
django-cors-headers==4.9.0
django-environ==0.13.0
django-filter==25.2
psycopg2-binary==2.9.11
stripe==15.0.0
Pillow==12.2.0
